refactor: fold dataset identity into evaluation
