+++ docs/guide/api-stability.md
@@ -38,23 +38,21 @@ experimental may change without the compatibility guarantees of the main API.
 
 ### Provisional capability packages
 
-`archetype.experiments`, `archetype.htn`, and `archetype.datasets` are not
-supported application surfaces merely because their modules are importable.
-They contain working prototypes and, in the dataset case, a normatively defined
-vocabulary, but their package boundaries may change before v1 as coding-mission
+`archetype.experiments` and `archetype.htn` are not supported application
+surfaces merely because their modules are importable. They contain working
+prototypes, but their package boundaries may change before v1 as coding-mission
 and physical-AI capabilities settle. The import-boundary checker constrains
 what these packages may depend on; it does not promote them to public API.
 
 The target ownership is recorded in
 [Agent Missions V1, section 9](agent-missions.md#9-family-direction-after-v1),
-but the orphan-package cleanup is not yet implemented. `archetype.htn` folds under
-`archetype.missions` as the planning subdomain via #591. `archetype.datasets`
-is removed by #590, which folds evidence identity into
-`archetype.evaluation.contracts`. The mixed `experiments` package is staged
-into `research` (#585), mission-owned trajectory modules (#586/#587), and
-`physical_ai` (#588/#589), then removed by #592. The surviving future family
-paths remain provisional until an implementation issue explicitly graduates a
-surface.
+and is landing incrementally. Dataset evidence identity now lives in
+`archetype.evaluation.contracts`, and the former `archetype.datasets` package is
+gone. `archetype.htn` folds under `archetype.missions` as the planning
+subdomain. The mixed `experiments` package is staged into research,
+mission-owned trajectory modules, and `physical_ai`, then removed. The
+surviving future family paths remain provisional until an implementation issue
+explicitly graduates a surface.
 
 Do not build a compatibility promise around those module paths yet. New
 applications use `ArchetypeRuntime` and supported extension types. A future
