+++ docs/guide/application-architecture.md
@@ -174,10 +174,11 @@ src/archetype/app/
 
 The mission-adjacent cleanup direction is recorded in
 [Agent Missions V1, section 9](agent-missions.md#9-family-direction-after-v1).
-It folds HTN planning and mission trajectory projections under missions,
-removes the datasets and experiments umbrellas, and gives physical AI a named
-family. Those moves are not prerequisites for V1; the current source tree and
-machine manifest remain implementation truth until each move lands.
+Dataset evidence identity has moved into evaluation and the datasets umbrella
+is gone. The remaining work folds HTN planning and mission trajectory
+projections under missions, removes the experiments umbrella, and gives
+physical AI a named family. The current source tree and machine manifest remain
+implementation truth until each move lands.
 
 Every application family co-locates its internal protocols, boundary models,
 and authority implementation. It imports reusable domain values from their
