--- quality/architecture.d/datasets.toml
@@ -1,8 +1,0 @@
-# Architecture policy fragment: datasets.
-# Merged into quality/architecture.toml by scripts/check_architecture.py;
-# fragments may declare only rule/exception arrays, never scalar policy.
-
-[[top_level_family_rule]]
-name = "datasets-domain-family"
-consumer = "archetype.datasets"
-allowed_families = []
