+++ quality/contracts.toml
@@ -313,7 +313,7 @@ source = "docs/guide/dataset-eval-ontology.md"
 section = "1. The contract in one view"
 owner = "evaluation"
 risk = "medium"
-pytest = ["tests/test_dataset_definitions.py"]
+pytest = ["tests/evaluation/test_dataset_identity_contracts.py"]
 static = []
 evals = ["spec.dataset_eval_ontology"]
 benchmarks = []
