+++ tests/evaluation/test_dataset_identity_contracts.py
@@ -1,7 +1,7 @@
 # Copyright 2026 Vangelis Technologies Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Executable contracts for the dataset/eval ontology vocabulary."""
+"""Executable contracts for evaluation-owned dataset evidence identity."""
 
 from __future__ import annotations
 
@@ -9,7 +9,7 @@ import dataclasses
 
 import pytest
 
-from archetype.datasets import (
+from archetype.evaluation import (
     EpisodeRef,
     Eval,
     Grader,
@@ -34,7 +34,7 @@ def _rubric() -> Rubric:
     )
 
 
-def test_eval_binds_one_task_to_a_non_empty_rubric():
+def test_eval_binds_one_task_to_a_non_empty_rubric() -> None:
     evaluation = Eval(task=_task(), rubric=_rubric())
 
     assert evaluation.task.key == ("libero", "libero_spatial", "3")
@@ -44,7 +44,7 @@ def test_eval_binds_one_task_to_a_non_empty_rubric():
     ]
 
 
-def test_trial_keeps_dataset_identity_and_runtime_provenance_separate():
+def test_trial_keeps_dataset_identity_and_runtime_provenance_separate() -> None:
     runtime = RuntimeSlice(
         world_id="world-7",
         run_id="run-9",
@@ -64,7 +64,7 @@ def test_trial_keeps_dataset_identity_and_runtime_provenance_separate():
     assert trial.episode.key == ("libero", 17)
 
 
-def test_reader_trial_may_have_no_runtime_provenance():
+def test_reader_trial_may_have_no_runtime_provenance() -> None:
     trial = Trial(
         task=_task(),
         seed=0,
@@ -75,13 +75,13 @@ def test_reader_trial_may_have_no_runtime_provenance():
 
 
 @pytest.mark.parametrize("episode_id", [-1, True, 1.5, "1"])
-def test_episode_id_is_a_non_negative_integer(episode_id):
+def test_episode_id_is_a_non_negative_integer(episode_id: object) -> None:
     error = ValueError if episode_id == -1 else TypeError
     with pytest.raises(error):
-        EpisodeRef(benchmark="libero", episode_id=episode_id)
+        EpisodeRef(benchmark="libero", episode_id=episode_id)  # type: ignore[arg-type]
 
 
-def test_trial_rejects_cross_benchmark_episode():
+def test_trial_rejects_cross_benchmark_episode() -> None:
     with pytest.raises(ValueError, match="same benchmark"):
         Trial(
             task=_task(),
@@ -90,7 +90,7 @@ def test_trial_rejects_cross_benchmark_episode():
         )
 
 
-def test_runtime_slice_rejects_reversed_tick_bounds():
+def test_runtime_slice_rejects_reversed_tick_bounds() -> None:
     with pytest.raises(ValueError, match="final_tick"):
         RuntimeSlice(
             world_id="world-7",
@@ -101,15 +101,16 @@ def test_runtime_slice_rejects_reversed_tick_bounds():
         )
 
 
-def test_rubric_requires_unique_graders():
+def test_rubric_requires_unique_graders() -> None:
     grader = Grader(name="success", kind=GraderKind.CHECK)
 
     with pytest.raises(ValueError, match="unique"):
         Rubric(graders=(grader, grader))
 
 
-def test_vocabulary_is_immutable():
+def test_vocabulary_is_immutable_and_evaluation_owned() -> None:
     task = _task()
 
+    assert TaskRef.__module__ == "archetype.evaluation.contracts"
     with pytest.raises(dataclasses.FrozenInstanceError):
         task.suite = "changed"  # type: ignore[misc]
